[feat] 좋아요한 모든 영화 조회 구현

+++ src/main/java/com/supershy/moviepedia/movie/entity/Movie.java
@@ -27,7 +27,7 @@ public class Movie {
 
     private String director;
 
-    private BigDecimal reservationRate;
+    private Double reservationRate;
 
     private String imageUrl;
